requirements.txt
Flask==2.1.1
pandas==1.4.2
joblib==1.1.0
scikit-learn==1.0.2
numpy==1.21.6
imblearn==0.0
Werkzeug==2.1.1
mysql-connector-python==8.0.33
python-dotenv==1.0.0
